Introduce DirectorySelectionActivity
